isl_multi_templ.c: extract out isl_multi_dims.c
This makes it easier to add functions that belong to the same category.
