+++ Makefile.am
@@ -372,6 +372,7 @@ EXTRA_DIST = \
 	isl_multi_apply_union_set.c \
 	isl_multi_cmp.c \
 	isl_multi_coalesce.c \
+	isl_multi_dims.c \
 	isl_multi_floor.c \
 	isl_multi_gist.c \
 	isl_multi_hash.c \
