+++ isl_val.c
@@ -1641,6 +1641,7 @@ isl_stat isl_val_check_match_domain_space(__isl_keep isl_val *v,
 #define NO_FROM_BASE
 #define NO_MOVE_DIMS
 #include <isl_multi_templ.c>
+#include <isl_multi_dims.c>
 
 /* Apply "fn" to each of the elements of "mv" with as second argument "v".
  */
